Resolve conflicts combining the electronic-excitation feature (elecmode/telec,
discrete electronic energy modes, eelec/temp_elec threading, icell-threaded
EEXCHANGE) with master's changes (gas/gas collision computes, tempscale-based
update_custom, post_run(), modernized Kokkos view accessors, and refactored
perform_collision control flow).

+++ src/comm.h
@@ -49,7 +49,7 @@ class Comm : protected Pointers {
  protected:
   class Irregular *iparticle,*igrid,*iuniform;
   char *sbuf,*rbuf;
-  int maxsendbuf,maxrecvbuf;
+  bigint maxsendbuf,maxrecvbuf;
   int *pproc,*gproc,*gsize;
   int maxpproc,maxgproc;
   bigint rvous_bytes;
